refactor group info widgets
