+++ lib/ui/chat/chat_info/group_chat_info.dart
@@ -262,458 +262,3 @@ class _GroupChatInfoState extends ConsumerState<GroupChatInfo> {
     );
   }
 }
-
-class GroupMemberBottomSheet extends ConsumerStatefulWidget {
-  const GroupMemberBottomSheet({
-    super.key,
-    required this.member,
-    required this.groupId,
-  });
-  final User member;
-  final String groupId;
-  static void show(BuildContext context, {required String groupId, required User member}) {
-    CustomBottomSheet.show(
-      context: context,
-      title: 'Member',
-
-      builder: (context) => GroupMemberBottomSheet(groupId: groupId, member: member),
-    );
-  }
-
-  @override
-  ConsumerState<GroupMemberBottomSheet> createState() => _GroupMemberBottomSheetState();
-}
-
-class _GroupMemberBottomSheetState extends ConsumerState<GroupMemberBottomSheet> {
-  String currentUserNpub = '';
-
-  void _copyToClipboard() {
-    final npub = widget.member.publicKey;
-    if (npub.isEmpty) {
-      ref.showErrorToast('No public key to copy');
-      return;
-    }
-    Clipboard.setData(ClipboardData(text: npub));
-    ref.showSuccessToast(
-      'Public Key copied.',
-    );
-  }
-
-  void _openAddToGroup() {
-    if (widget.member.publicKey.isEmpty) {
-      ref.showErrorToast('No user to add to group');
-      return;
-    }
-    context.push('/add_to_group/${widget.member.publicKey}');
-  }
-
-  void _loadCurrentUserNpub() async {
-    final activeAccount = ref.read(activeAccountProvider);
-    if (activeAccount != null) {
-      currentUserNpub = await activeAccount.toNpub() ?? '';
-      setState(() {});
-    }
-  }
-
-  void _openRemoveFromGroupDialog() {
-    showDialog(
-      context: context,
-      builder: (context) {
-        return WhitenoiseDialog.custom(
-          customChild: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Text(
-                'Are you sure you want to remove ${widget.member.name} from the group?',
-                style: context.textTheme.bodyMedium?.copyWith(
-                  color: context.colors.primary,
-                  fontSize: 16.sp,
-                ),
-              ),
-              Gap(16.h),
-              AppFilledButton.child(
-                onPressed: () {},
-                visualState: AppButtonVisualState.secondaryWarning,
-                child: Text('Remove'),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _loadCurrentUserNpub();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final currentUserIsAdmin =
-        ref
-            .watch(groupsProvider)
-            .groupAdmins?[widget.groupId]
-            ?.firstWhereOrNull(
-              (admin) => admin.publicKey == currentUserNpub,
-            ) !=
-        null;
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Gap(16.h),
-        ContactAvatar(
-          imageUrl: widget.member.imagePath ?? '',
-          displayName: widget.member.username,
-          size: 96.w,
-        ),
-        Gap(4.h),
-        Text(
-          widget.member.username ?? '',
-          style: TextStyle(
-            fontWeight: FontWeight.w600,
-            color: context.colors.primary,
-          ),
-          textAlign: TextAlign.center,
-        ),
-        if (widget.member.nip05.isNotEmpty)
-          Text(
-            widget.member.nip05,
-            style: TextStyle(
-              color: context.colors.mutedForeground,
-            ),
-            textAlign: TextAlign.center,
-          ),
-        Gap(16.h),
-        Padding(
-          padding: EdgeInsets.symmetric(horizontal: 8.w),
-          child: Row(
-            children: [
-              Flexible(
-                child: Text(
-                  widget.member.publicKey.formatPublicKey(),
-                  textAlign: TextAlign.center,
-                  style: context.textTheme.bodyMedium?.copyWith(
-                    color: context.colors.mutedForeground,
-                    fontSize: 14.sp,
-                  ),
-                ),
-              ),
-              Gap(8.w),
-              InkWell(
-                onTap: _copyToClipboard,
-                child: SvgPicture.asset(
-                  AssetsPaths.icCopy,
-                  width: 24.w,
-                  height: 24.w,
-                  colorFilter: ColorFilter.mode(
-                    context.colors.primary,
-                    BlendMode.srcIn,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        Gap(32.h),
-        if (currentUserIsAdmin)
-          Column(
-            children: [
-              Row(
-                spacing: 6.w,
-                children: [
-                  Flexible(
-                    child: _SendMessageButton(widget.member),
-                  ),
-                  Flexible(
-                    child: _AddToContactButton(widget.member),
-                  ),
-                ],
-              ),
-              // Gap(8.h),
-              // AppFilledButton.child(
-              //   onPressed: () {},
-              //   size: AppButtonSize.small,
-              //   visualState: AppButtonVisualState.secondary,
-              //   child: Row(
-              //     mainAxisAlignment: MainAxisAlignment.center,
-              //     children: [
-              //       Text(
-              //         'Make Admin',
-              //         style: context.textTheme.bodyMedium?.copyWith(
-              //           color: context.colors.primary,
-              //           fontWeight: FontWeight.w600,
-              //           fontSize: 14.sp,
-              //         ),
-              //       ),
-              //       Gap(8.w),
-              //       SvgPicture.asset(
-              //         AssetsPaths.icMakeAdmin,
-              //         width: 14.w,
-              //         height: 13.h,
-              //         colorFilter: ColorFilter.mode(
-              //           context.colors.primary,
-              //           BlendMode.srcIn,
-              //         ),
-              //       ),
-              //     ],
-              //   ),
-              // ),
-            ],
-          )
-        else
-          _SendMessageButton(widget.member),
-        Gap(8.h),
-        AppFilledButton.child(
-          onPressed: _openAddToGroup,
-          size: AppButtonSize.small,
-          visualState: AppButtonVisualState.secondary,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Text(
-                'Add to Another Group',
-                style: context.textTheme.bodyMedium?.copyWith(
-                  color: context.colors.primary,
-                  fontWeight: FontWeight.w600,
-                  fontSize: 14.sp,
-                ),
-              ),
-              Gap(8.w),
-              SvgPicture.asset(
-                AssetsPaths.icChatInvite,
-                width: 14.w,
-                height: 13.h,
-                colorFilter: ColorFilter.mode(
-                  context.colors.primary,
-                  BlendMode.srcIn,
-                ),
-              ),
-            ],
-          ),
-        ),
-        if (currentUserIsAdmin) ...[
-          Gap(8.h),
-          AppFilledButton.child(
-            onPressed: _openRemoveFromGroupDialog,
-            size: AppButtonSize.small,
-            visualState: AppButtonVisualState.secondaryWarning,
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Text(
-                  'Remove From Group',
-                  style: context.textTheme.bodyMedium?.copyWith(
-                    color: context.colors.destructive,
-                    fontWeight: FontWeight.w600,
-                    fontSize: 14.sp,
-                  ),
-                ),
-                Gap(8.w),
-                SvgPicture.asset(
-                  AssetsPaths.icRemoveOutlined,
-                  width: 14.w,
-                  height: 13.h,
-                  colorFilter: ColorFilter.mode(
-                    context.colors.destructive,
-                    BlendMode.srcIn,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ] else ...[
-          Gap(8.h),
-          _AddToContactButton(widget.member),
-        ],
-      ],
-    );
-  }
-}
-
-class _SendMessageButton extends ConsumerStatefulWidget {
-  const _SendMessageButton(this.user);
-  final User user;
-  @override
-  ConsumerState<_SendMessageButton> createState() => __SendMessageButtonState();
-}
-
-class __SendMessageButtonState extends ConsumerState<_SendMessageButton> {
-  final _logger = Logger('_SendMessageButtonState');
-  bool _isCreatingGroup = false;
-  bool get _isLoading => _isCreatingGroup;
-
-  Future<void> _createOrOpenDirectMessageGroup() async {
-    if (widget.user.publicKey.isEmpty) {
-      ref.showErrorToast('No user to start chat with');
-      return;
-    }
-    setState(() {
-      _isCreatingGroup = true;
-    });
-
-    try {
-      final groupData = await ref
-          .read(groupsProvider.notifier)
-          .createNewGroup(
-            groupName: 'DM',
-            groupDescription: 'Direct message',
-            memberPublicKeyHexs: [widget.user.publicKey],
-            adminPublicKeyHexs: [widget.user.publicKey],
-          );
-
-      if (groupData != null) {
-        _logger.info('Direct message group created successfully: ${groupData.mlsGroupId}');
-
-        if (mounted) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            if (mounted) {
-              Navigator.pop(context);
-              context.navigateToGroupChatAndPopToHome(groupData);
-            }
-          });
-
-          ref.showSuccessToast(
-            'Chat with ${widget.user.username ?? widget.user.name} started successfully',
-          );
-        }
-      } else {
-        // Group creation failed - check the provider state for the error message
-        if (mounted) {
-          final groupsState = ref.read(groupsProvider);
-          final errorMessage = groupsState.error ?? 'Failed to create direct message group';
-          ref.showErrorToast(errorMessage);
-        }
-      }
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isCreatingGroup = false;
-        });
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AppFilledButton.child(
-      onPressed: _createOrOpenDirectMessageGroup,
-      loading: _isLoading,
-      size: AppButtonSize.small,
-      visualState: AppButtonVisualState.secondary,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(
-            'Send Message',
-            style: context.textTheme.bodyMedium?.copyWith(
-              color: context.colors.primary,
-              fontWeight: FontWeight.w600,
-              fontSize: 14.sp,
-            ),
-          ),
-          Gap(8.w),
-          SvgPicture.asset(
-            AssetsPaths.icMessage,
-            width: 14.w,
-            height: 13.h,
-            colorFilter: ColorFilter.mode(
-              context.colors.primary,
-              BlendMode.srcIn,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// add contact
-class _AddToContactButton extends ConsumerStatefulWidget {
-  const _AddToContactButton(this.user);
-  final User user;
-  @override
-  ConsumerState<_AddToContactButton> createState() => __AddToContactButtonState();
-}
-
-class __AddToContactButtonState extends ConsumerState<_AddToContactButton> {
-  bool _isAddingContact = false;
-  bool get _isLoading => _isAddingContact;
-  bool _isContact() {
-    final contactsState = ref.watch(contactsProvider);
-    final contacts = contactsState.contactModels ?? [];
-
-    // Check if the current user's pubkey exists in contacts
-    return contacts.any(
-      (contact) => contact.publicKey.toLowerCase() == widget.user.publicKey.toLowerCase(),
-    );
-  }
-
-  Future<void> _toggleContact() async {
-    setState(() {
-      _isAddingContact = true;
-    });
-
-    try {
-      final contactsNotifier = ref.read(contactsProvider.notifier);
-      final isCurrentlyContact = _isContact();
-
-      if (isCurrentlyContact) {
-        await contactsNotifier.removeContactByHex(widget.user.publicKey);
-        if (mounted) {
-          ref.showSuccessToast('${widget.user.name} removed from contacts');
-        }
-      } else {
-        await contactsNotifier.addContactByHex(widget.user.publicKey);
-        if (mounted) {
-          ref.showSuccessToast('${widget.user.name} added to contacts');
-        }
-      }
-    } catch (e) {
-      if (mounted) {
-        ref.showErrorToast('Failed to update contact: $e');
-      }
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isAddingContact = false;
-        });
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isContact = _isContact();
-    return AppFilledButton.child(
-      onPressed: _toggleContact,
-      loading: _isLoading,
-      size: AppButtonSize.small,
-      visualState: AppButtonVisualState.secondary,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(
-            isContact ? 'Remove Contact' : 'Add Contact',
-            style: context.textTheme.bodyMedium?.copyWith(
-              color: context.colors.primary,
-              fontWeight: FontWeight.w600,
-              fontSize: 14.sp,
-            ),
-          ),
-          Gap(8.w),
-          SvgPicture.asset(
-            isContact ? AssetsPaths.icRemoveUser : AssetsPaths.icAddUser,
-            width: 11.w,
-            height: 11.w,
-            colorFilter: ColorFilter.mode(
-              context.colors.primary,
-              BlendMode.srcIn,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
